Treat releaseNightly same as nightly
Gradle introduced a separate releaseNightly property to distinguish
nightly builds from the release branch from the ones on the master
branch. This broke our filtering, so we need to account for it.

This is not the ideal solution, since we're still dependent on Gradle
not adding another similar property to identify another kind of
pre-release version.

This fixes #18.