Filter out outdated patches
Previously, we would include all patch versions for a minor when you
weren't in sparse mode. These aren't often worth testing, since they are
intended to be only compatible fixes to the minor.

This also now allows for including rcs or milestones that happen within
ranges and ensures that the most mature version for a given minor is the
only one included (except in the case where an rc exists for a patch).

This fixes #17.