Support Java toolchains
Now that Gradle's Java toolchain feature is available, it means we can
run Gradle under a different version than we run tasks against. In our
case, that means we can generate tasks for every Java version and Gradle
version combination, without requiring you to externally do a CI matrix
build.

This does involve a lot of breaking changes to the interface as part of
the process, but they allow us to support this new feature and switch to
approaches that take advantage of task creation avoidance, configuration
cache support, etc.