Eagerly process matrices to register tasks
While tasks are still created lazily, we now eagerly resolve the
matrices so that we can register the tasks. Without this, if someone
uses register instead of create in their buildscript to create the
matrix, we never generate tasks, so they aren't available as options to
run.

This should still be in line with Gradle best practices since tasks are
still registered instead of created, so task avoidance is still
maintained.